+++ src/Connections/Connection.php
@@ -2,9 +2,9 @@
 namespace Michaels\Spider\Connections;
 
 use Michaels\Manager\Traits\ManagesItemsTrait;
+use Michaels\Spider\Commands\CommandInterface;
 use Michaels\Spider\Drivers\DriverInterface;
 use Michaels\Spider\Graphs\Graph;
-use Michaels\Spider\Queries\CommandInterface;
 
 /**
  * Facilitates two-way communication with a driver store